requirements.txt
streamlit==1.27.2
torch==2.1.2
torchvision==0.16.2
opencv-python==4.8.0.76
Pillow==9.5.0
numpy==1.24.4
gdown==5.1.0
realesrgan==0.3.0
